fix(home): restore configurable meta accent stripe
